Fix ETA type to Long

+++ src/main/java/com/softdev/softdev/dto/busstop/BusStopETADTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class BusStopETADTO {
     private Long bus_id;
     private Long stop_id;
-    private double eta_seconds;
+    private Long eta_seconds;
 }
